+++ app/app.go
@@ -36,7 +36,7 @@ import (
 	"github.com/sge-network/sge/app/keepers"
 	sgeappparams "github.com/sge-network/sge/app/params"
 	"github.com/sge-network/sge/app/upgrades"
-	v1 "github.com/sge-network/sge/app/upgrades/v1"
+	v110 "github.com/sge-network/sge/app/upgrades/v110"
 	"github.com/spf13/cast"
 	abci "github.com/tendermint/tendermint/abci/types"
 	tmjson "github.com/tendermint/tendermint/libs/json"
@@ -66,7 +66,9 @@ func getGovProposalHandlers() []govclient.ProposalHandler {
 var (
 	// DefaultNodeHome default home directories for the application daemon
 	DefaultNodeHome string
-	Upgrades        = []upgrades.Upgrade{v1.Upgrade}
+	Upgrades        = []upgrades.Upgrade{
+		v110.Upgrade,
+	}
 )
 
 var (
