+++ app/upgrades/v110/upgrades.go
@@ -1,4 +1,4 @@
-package v1
+package v110
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
